feat: try worker
