+++ .env
@@ -16,6 +16,7 @@
 
 ###> symfony/framework-bundle ###
 APP_ENV=dev
+APP_DEBUG=1
 APP_SECRET=ad88d8c60b8d8bb76fdb1ea46fe98365
 ###< symfony/framework-bundle ###
 
@@ -31,3 +32,10 @@ DATABASE_URL="postgresql://yousign:123456789@database:5432/gh-archive-keyword?se
 
 DOWNLOAD_DIRECTORY=%kernel.project_dir%/var/downloads
 EXTRACTION_DIRECTORY=%kernel.project_dir%/var/downloads/extraction
+
+###> symfony/messenger ###
+# Choose one of the transports below
+# MESSENGER_TRANSPORT_DSN=amqp://guest:guest@localhost:5672/%2f/messages
+MESSENGER_TRANSPORT_DSN=redis://redis:6379/messages
+# MESSENGER_TRANSPORT_DSN=doctrine://default?auto_setup=0
+###< symfony/messenger ###
