+++ composer.json
@@ -19,10 +19,12 @@
         "symfony/flex": "^2.4.7",
         "symfony/framework-bundle": "^6.4.12",
         "symfony/http-client": "^6.4",
+        "symfony/messenger": "^6.4",
         "symfony/monolog-bundle": "^3.10",
         "symfony/property-access": "^6.4.11",
         "symfony/property-info": "^6.4.10",
         "symfony/proxy-manager-bridge": "^6.4.8",
+        "symfony/redis-messenger": "^6.4",
         "symfony/runtime": "^6.4.12",
         "symfony/serializer": "^6.4.12",
         "symfony/validator": "^6.4.12",
